+++ extensions/toggl-track/src/index.tsx
@@ -1,11 +1,13 @@
-import { ActionPanel, Icon, List, Action } from "@raycast/api";
+import { Action, ActionPanel, Alert, Color, confirmAlert, Icon, List } from "@raycast/api";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 
+import { removeTimeEntry } from "@/api/timeEntries";
 import TimeEntryForm from "@/components/CreateTimeEntryForm";
 import RunningTimeEntry from "@/components/RunningTimeEntry";
 import { ExtensionContextProvider } from "@/context/ExtensionContext";
 import { formatSeconds } from "@/helpers/formatSeconds";
+import { Verb, withToast } from "@/helpers/withToast";
 import { useProcessedTimeEntries } from "@/hooks/useProcessedTimeEntries";
 import { useTimeEntryActions } from "@/hooks/useTimeEntryActions";
 import { useTotalDurationToday } from "@/hooks/useTotalDurationToday";
@@ -14,11 +16,12 @@ dayjs.extend(duration);
 
 function ListView() {
   const {
-    timeEntries,
-    runningTimeEntry,
     isLoading,
-    revalidateTimeEntries,
+    mutateTimeEntries,
     revalidateRunningTimeEntry,
+    revalidateTimeEntries,
+    runningTimeEntry,
+    timeEntries,
     timeEntriesWithUniqueProjectAndDescription,
   } = useProcessedTimeEntries();
 
@@ -68,7 +71,10 @@ function ListView() {
               keywords={[timeEntry.description, timeEntry.project_name || "", timeEntry.client_name || ""]}
               title={timeEntry.description || "No description"}
               subtitle={(timeEntry.client_name ? timeEntry.client_name + " | " : "") + (timeEntry.project_name ?? "")}
-              accessories={[...timeEntry.tags.map((tag) => ({ tag })), { text: timeEntry.billable ? "$" : "" }]}
+              accessories={[
+                ...timeEntry.tags.map((tag) => ({ tag })),
+                timeEntry.billable ? { tag: { value: "$" } } : {},
+              ]}
               icon={{ source: Icon.Circle, tintColor: timeEntry.project_color }}
               actions={
                 <ActionPanel>
@@ -90,6 +96,37 @@ function ListView() {
                       </ExtensionContextProvider>
                     }
                   />
+                  <ActionPanel.Section>
+                    <Action
+                      title="Delete Time Entry"
+                      icon={Icon.Trash}
+                      style={Action.Style.Destructive}
+                      shortcut={{ modifiers: ["ctrl"], key: "x" }}
+                      onAction={async () => {
+                        await confirmAlert({
+                          title: "Delete Time Entry",
+                          message: "Are you sure you want to delete this time entry?",
+                          icon: {
+                            source: Icon.Trash,
+                            tintColor: Color.Red,
+                          },
+                          primaryAction: {
+                            title: "Delete",
+                            style: Alert.ActionStyle.Destructive,
+                            onAction: () => {
+                              withToast({
+                                noun: "Time Entry",
+                                verb: Verb.Delete,
+                                action: async () => {
+                                  await mutateTimeEntries(removeTimeEntry(timeEntry.workspace_id, timeEntry.id));
+                                },
+                              });
+                            },
+                          },
+                        });
+                      }}
+                    />
+                  </ActionPanel.Section>
                 </ActionPanel>
               }
             />